backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
python-dotenv==1.0.0

# IA y Machine Learning - PyTorch se instala en Dockerfile
ultralytics==8.0.200
opencv-python-headless==4.8.1.78
numpy==1.24.3

# APIs externas - OpenAI ACTUALIZADO
openai==1.66.3
google-cloud-speech==2.21.0
google-auth==2.23.4
firebase-admin==6.2.0
requests==2.31.0

# Base de datos y producción
psycopg2-binary==2.9.7
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# Cache y Redis
redis==4.5.5
django-redis==5.3.0

# Procesamiento de imágenes y multimedia
Pillow==10.1.0